chore(sdk): move action proxy in proxy folder
