+++ packages/sdk/src/plugin/proxies/index.ts
@@ -1,0 +1,1 @@
+export * as action from './action'
